Change fileClosed behavior
