+++ src/main/java/io/openliberty/tools/intellij/lsp4mp/lsp4ij/ConnectDocumentToLanguageServerSetupParticipant.java
@@ -16,6 +16,7 @@ import com.intellij.openapi.fileEditor.FileDocumentManager;
 import com.intellij.openapi.fileEditor.FileEditorManager;
 import com.intellij.openapi.fileEditor.FileEditorManagerListener;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.project.ProjectManagerListener;
 import com.intellij.openapi.vfs.VirtualFile;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
@@ -48,28 +49,6 @@ public class ConnectDocumentToLanguageServerSetupParticipant implements ProjectC
             // Force the start of all languages servers mapped with the given file
             LanguageServiceAccessor.getInstance(source.getProject())
                     .getLanguageServers(document, capabilities -> true);
-        }
-    }
-
-    @Override
-    public void fileClosed(@NotNull FileEditorManager source, @NotNull VirtualFile file) {
-        URI uri = LSPIJUtils.toUri(file);
-        if (uri != null) {
-            try {                
-                // TODO: revisit this code, because it can restart language servers
-                // when a diagnostics is published after the file is closed and the project is closed
-                // See https://github.com/redhat-developer/intellij-quarkus/issues/840
-                // Remove the cached file wrapper if needed
-                LSPVirtualFileWrapper.dispose(file);
-                // Disconnect the given file from all language servers
-                LanguageServiceAccessor.getInstance(source.getProject())
-                        .getLSWrappers(file, capabilities -> true)
-                        .forEach(
-                                wrapper -> wrapper.disconnect(uri)
-                        );
-            } catch (Exception e) {
-                LOGGER.warn("Error while disconnecting the file '" + uri + "' from all language servers", e);
-            }
         }
     }
