+++ src/main/java/io/openliberty/tools/intellij/lsp4mp/lsp4ij/LanguageClientImpl.java
@@ -125,4 +125,8 @@ public class LanguageClientImpl implements LanguageClient {
         return future;
     }
 
+    // public void dispose() {
+
+    // }
+
 }
